requirements.txt
click
Flask
itsdangerous
Jinja2
MarkupSafe
mysql-connector
Werkzeug
gunicorn
certifi
chardet
configparser
idna
lxml
pip
requests
setuptools
urllib3
wheel
pylint